requirements.txt
# Core Dependencies for Neural Additive Model (NAM)
# Python 3.8+ required

# Deep Learning Framework
tensorflow>=2.13.0
keras>=2.13.0

# Data Processing
pandas>=1.5.0
numpy>=1.23.0

# Machine Learning
scikit-learn>=1.2.0
statsmodels>=0.13.5

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.2.0

# Optimization
scipy>=1.9.0
optuna>=3.1.0  # For hyperparameter tuning

# Dashboard (optional)
streamlit>=1.20.0  # For interactive dashboard

# Additional ML Tools
xgboost>=1.7.0  # For baseline models
lightgbm>=3.3.0  # For baseline models

# Data Validation
great-expectations>=0.16.0  # Optional, for data validation

# Testing (optional)
pytest>=7.2.0
pytest-cov>=4.0.0

# Code Quality (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0